apitest3/requirement.txt
django
mysqlclient
djangorestframework
drf-spectacular
djangorestframework-simplejwt
pillow